Migrate TPU orchestration from levanter to fray
Move Ray TPU orchestration infrastructure (~1600 lines) from levanter to fray for reusability across frameworks. This enables generic cluster abstractions to support TPU gang scheduling.

**Migration Details:**
- Created `fray/cluster/ray/tpu/` module with config, types, orchestration, execution, and utils
- Migrated 39 TPU configurations (v4, v5litepod, v5p, v6e)
- Migrated actor hierarchy: ResourcePoolManager, SlicePoolManager, SliceActor, TPUHostActor
- Migrated execution functions: run_on_pod variants with retry/preemption logic
- Migrated 480 lines of tests to `fray/tests/ray/test_tpu.py`
- Updated 7 consumer imports in marin and scripts
- Created backward compatibility wrapper in levanter re-exporting from fray
- Docker-related functions remain in levanter as levanter-specific infrastructure

**API Changes:**
- Import paths changed from `levanter.infra.ray_tpu` to `fray.cluster.ray.tpu`
- Backward compatibility maintained via re-exports in levanter

**Testing:**
- All imports verified working (both fray and levanter paths)
- Test file migrated with proper imports
- Linting/formatting applied

Note: 17 E501 line-length violations in orchestration.py logging messages are acceptable.

Addresses #2003
